split out _metadata_wrapper.py from _tdb_handles.py
